Create AnalyticsProvider to propogate amplitude functionality across app
