--- packages/synapse-interface/components/analytics/index.ts
@@ -1,1 +1,0 @@
-import * as amplitude from '@amplitude/analytics-browser'
